Add tls-verify options to docker commands (for podman support)
Because many of the installations of galaxy_ng we use for testing will
use self-signed certs, if we want to use podman in our tests we need to
make it possible to turn off tls-verify.

These changes make that possible.

In addition this commit simplifies the interface for the  `docker tag`
argument - now rather than requiring an explicit `version` argument, the
caller of the function should simply put the version in the tag that
they applying to the image.